Update sketch_2018_01_24.ino
Slightly error prone as someone might change the password length and forget to change password_length. Usually it is done like this. Sizeof returns the total ammount of bytes in memory, so we need to divide by the ammount of memory one item of the array takes to get the ammount of items. So usually you would use: (sizeof(_Array) / sizeof(_Array[0]))
Since typing this gets annoying, you would usually define a macro to do that for you, done in like 11 so you can simply use:
Another way would be to always use "_countof(password_array)" instead of "password_length). But this is more a question of preferrence.

On a general note, I would try not to mix configuration values like password_array and derived values like password_length. Just have a block of all configuration values that could be changed at the top and simple variable initialisations (eg. input_pos = 0) further down in their own block.